skills/unistore-query-analyzer/requirements.txt
numpy
pandas
sqlglot
snowflake-snowpark-python
